+++ src/app/models/user/user-preferences.mock.ts
@@ -1,0 +1,11 @@
+import { Mock } from '../mock';
+import { UserPreferences } from './user-preferences';
+
+export class UserPreferencesMock extends Mock<UserPreferences> {
+    public mockOne(): UserPreferences {
+        const tmp = new UserPreferences();
+        tmp.killchain = 'TacticXYZ';
+        return tmp;
+    }
+}
+export const UserPreferencesMockFactory = new UserPreferencesMock();
